backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
PyPDF2>=3.0.1
python-multipart>=0.0.9
pydantic>=2.6.0
jose>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.4.4
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
pytz>=2024.1
